requirements.txt
numpy
tqdm
pillow
opencv-python
